Fix version number.
Note to maintainer: Please do NOT push changes with the 'git merge'
headers in them. This is not the right way to do things. IF the merge is
accepted once, you will NOT be able to push again as the pre-receive
hook will prevent it.

If you have any questions, please email nitesh.turaga@roswellpark.org